+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [2.0.12](https://github.com/phalcon/cphalcon/releases/tag/phalcon-v2.0.12) (????-??-??)
+- Fixed regression changes for `Phalcon\Mvc\View\Engine\Volt::callMacro` [#11745](https://github.com/phalcon/cphalcon/issues/11745)
 
 # [2.0.11](https://github.com/phalcon/cphalcon/releases/tag/phalcon-v2.0.11) (2016-05-04)
 - Fix Model magic set functionality to maintain variable visibility and utilize setter methods.[#11286](https://github.com/phalcon/cphalcon/issues/11286)